+++ packages/hydrogen/src/components/ProductProvider/index.ts
@@ -3,3 +3,4 @@ export {
   ProductProviderFragment,
 } from './ProductProvider.client';
 export {useProduct} from '../../hooks/useProduct/useProduct';
+export type {Product} from './types';