+++ packages/hydrogen/src/graphql/graphql-constants.ts
@@ -2207,8 +2207,52 @@ fragment Model3DFragment on Model3d {
  *         ...ImageFragment
  *       }
  *     }
+ *     ... on Product {
+ *       handle
+ *       id
+ *       title
+ *       compareAtPriceRange {
+ *         maxVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *         minVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *       }
+ *       priceRange {
+ *         maxVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *         minVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *       }
+ *       variants (first: 1) {
+ *         edges {
+ *           node {
+ *             id
+ *             title
+ *             availableForSale
+ *             image {
+ *               ...ImageFragment
+ *             }
+ *             priceV2 {
+ *               ...MoneyFragment
+ *             }
+ *             compareAtPriceV2 {
+ *               ...MoneyFragment
+ *             }
+ *             selectedOptions {
+ *               name
+ *               value
+ *             }
+ *           }
+ *         }
+ *       }
+ *     }
  *   }
  * }
+ *
  * fragment ImageFragment on Image {
  *   id
  *   url
@@ -2238,8 +2282,52 @@ fragment MetafieldFragment on Metafield {
         ...ImageFragment
       }
     }
+    ... on Product {
+      handle
+      id
+      title
+      compareAtPriceRange {
+        maxVariantPrice {
+          ...MoneyFragment
+        }
+        minVariantPrice {
+          ...MoneyFragment
+        }
+      }
+      priceRange {
+        maxVariantPrice {
+          ...MoneyFragment
+        }
+        minVariantPrice {
+          ...MoneyFragment
+        }
+      }
+      variants (first: 1) {
+        edges {
+          node {
+            id
+            title
+            availableForSale
+            image {
+              ...ImageFragment
+            }
+            priceV2 {
+              ...MoneyFragment
+            }
+            compareAtPriceV2 {
+              ...MoneyFragment
+            }
+            selectedOptions {
+              name
+              value
+            }
+          }
+        }
+      }
+    }
   }
 }
+
 fragment ImageFragment on Image {
   id
   url
@@ -2390,8 +2478,52 @@ export const MoneyFragment: string = `fragment MoneyFragment on MoneyV2 {
  *         ...ImageFragment
  *       }
  *     }
+ *     ... on Product {
+ *       handle
+ *       id
+ *       title
+ *       compareAtPriceRange {
+ *         maxVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *         minVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *       }
+ *       priceRange {
+ *         maxVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *         minVariantPrice {
+ *           ...MoneyFragment
+ *         }
+ *       }
+ *       variants (first: 1) {
+ *         edges {
+ *           node {
+ *             id
+ *             title
+ *             availableForSale
+ *             image {
+ *               ...ImageFragment
+ *             }
+ *             priceV2 {
+ *               ...MoneyFragment
+ *             }
+ *             compareAtPriceV2 {
+ *               ...MoneyFragment
+ *             }
+ *             selectedOptions {
+ *               name
+ *               value
+ *             }
+ *           }
+ *         }
+ *       }
+ *     }
  *   }
  * }
+ *
  *
  * fragment VariantFragment on ProductVariant {
  *   id
@@ -2698,8 +2830,52 @@ fragment MetafieldFragment on Metafield {
         ...ImageFragment
       }
     }
+    ... on Product {
+      handle
+      id
+      title
+      compareAtPriceRange {
+        maxVariantPrice {
+          ...MoneyFragment
+        }
+        minVariantPrice {
+          ...MoneyFragment
+        }
+      }
+      priceRange {
+        maxVariantPrice {
+          ...MoneyFragment
+        }
+        minVariantPrice {
+          ...MoneyFragment
+        }
+      }
+      variants (first: 1) {
+        edges {
+          node {
+            id
+            title
+            availableForSale
+            image {
+              ...ImageFragment
+            }
+            priceV2 {
+              ...MoneyFragment
+            }
+            compareAtPriceV2 {
+              ...MoneyFragment
+            }
+            selectedOptions {
+              name
+              value
+            }
+          }
+        }
+      }
+    }
   }
 }
+
 
 fragment VariantFragment on ProductVariant {
   id